requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
pyyaml>=6.0
requests>=2.28.2
numpy>=1.24.2
tqdm>=4.65.0
python-multipart>=0.0.6
